fixed helper functions

+++ include/PA02/HelperFunctions.h
@@ -12,7 +12,7 @@ namespace Helper {
     void remapValues(ImageType & img);
 
     //function that applies mask to image at location specified by row and col and returns the new value. The maskCenterRow and maskCenterCol specify the center of the mask.
-    int applyMask(const ImageType & img, const ImageType & mask, int row, int col, unsigned int maskCenterRow, unsigned int maskCenterCol, bool normalizeMask = true);
+    double applyMask(const ImageType & img, const ImageType & mask, int row, int col, unsigned int maskCenterRow, unsigned int maskCenterCol, bool normalizeMask = true);
 
 };
